requirements.txt
aiohttp==3.5.4
aiohttp-cors==0.7.0
aioredis==1.2.0
amqp==2.5.0
argh==0.26.2
asn1crypto==0.24.0
astroid==2.2.5
async-timeout==3.0.1
asynctest==0.13.0
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
bcrypt==3.1.7
billiard==3.6.0.0
celery==4.3.0
Cerberus==1.3.1
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.4
cryptography==2.7
devtools==0.5
docopt==0.6.2
expects==0.9.0
glob2==0.7
hiredis==1.0.0
idna==2.8
importlib-metadata==0.18
isort==4.3.21
kombu==4.6.3
lazy-object-proxy==1.4.1
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
mock==3.0.5
more-itertools==7.1.0
motor==2.0.0
multidict==4.5.2
mypy==0.770
mypy-extensions==0.4.3
packaging==19.0
parse==1.12.1
parse-type==0.5.2
password-strength==0.0.3.post2
pathtools==0.1.2
pluggy==0.12.0
py==1.8.0
pycodestyle==2.5.0
pycparser==2.19
pydash==4.7.5
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.8.0
pyparsing==2.4.0
pytest==5.4.1
pytest-asyncio==0.10.0
pytest-cov==2.7.1
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.1
redis==3.2.1
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.3
vine==1.3.0
watchdog==0.9.0
watchgod==0.4
wcwidth==0.1.7
wrapt==1.11.2
yarl==1.3.0
zipp==0.5.2
